just fixed some ui

+++ routes/web.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\ImageController;
 use App\Http\Controllers\VisitorPageController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\ReservationController;
+use App\Http\Controllers\SubmitReportController;
 use App\Http\Controllers\RenterController;
 use Illuminate\Support\Facades\DB;
 
@@ -34,42 +35,39 @@ use Illuminate\Support\Facades\Route;
 // specify landing page 
 Route::middleware(['auth','verified'])->group( function(){
     Route::group(['middleware' => ['isAdmin']], function () {
-        Route::get('/dashboard',[AdminDashboardController::class, 'index'])->name('dashboard');
-        Route::get('/users', [UserController::class, 'index'])->name('users.index');
+        Route::get('/admin/dashboard',[AdminDashboardController::class, 'index'])->name('dashboard');
+        Route::get('/admin/users', [UserController::class, 'index'])->name('admin.users.index');
 
         // controller group for category
         Route::controller(CategoryController::class)->group(function() {
-            Route::get('/categories/index', 'index')->name('categories.index');
-            Route::post('/categories/index', 'create')->name('categories.create');
-            Route::get('/categories/{categories}/edit', 'edit')->name('categories.edit');
-            Route::put('/categories/{categories}/update', 'update')->name('categories.update');
-            Route::delete('/categories/{categories}/delete', 'delete')->name('categories.delete');
-            Route::get('/categories/restore','restore')->name('categories.restore');
+            Route::get('/admin/categories/index', 'index')->name('admin.categories.index');
+            Route::post('/admin/categories/index', 'create')->name('categories.create');
+            Route::get('/admin/categories/{categories}/edit', 'edit')->name('categories.edit');
+            Route::put('/admin/categories/{categories}/update', 'update')->name('categories.update');
+            Route::delete('/admin/categories/{categories}/delete', 'delete')->name('categories.delete');
+            Route::get('/admin/categories/restore','restore')->name('categories.restore');
         });
         // controller group for report
         Route::controller(ReportController::class)->group(function() {
-            Route::post('/reports/create','create')->name('reports.create');
-            Route::get('/reports', 'index')->name('reports.index');
+            Route::post('/admin/reports/create','create')->name('reports.create');
+            Route::get('/admin/reports', 'index')->name('admin.reports.index');
         });
         // controller group for category
         Route::controller(AppartmentController::class)->group(function() {
-            Route::get('/appartment', 'index')->name('appartment.index');
-            Route::post('/appartment/create', 'create')->name('appartment.create');
-            Route::get('/appartment/{apartment}/edit', 'edit')->name('appartment.edit');
-            Route::put('/appartment/{apartment}/update', 'update')->name('appartment.update');
-            Route::delete('/appartment/{apartment}/delete', 'delete')->name('appartment.delete');
+            Route::get('/admin/appartment', 'index')->name('admin.apartment.index');
+            Route::post('/admin/appartment/create', 'create')->name('appartment.create');
+            Route::get('/admin/appartment/{apartment}/edit', 'edit')->name('appartment.edit');
+            Route::put('/admin/appartment/{apartment}/update', 'update')->name('appartment.update');
+            Route::delete('/admin/appartment/{apartment}/delete', 'delete')->name('appartment.delete');
         });
         Route::controller(ImageController::class)->group(function() {
-            Route::get('/categories/{categoryId}/upload', 'index');
-            Route::post('/categories/{categoryId}/upload', 'store');
-            Route::get('/category-image/{categoryImageId}/delete', 'delete');
-
-            Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-            Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-            Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+            Route::get('/admin/categories/{categoryId}/upload', 'index');
+            Route::post('/admin/categories/{categoryId}/upload', 'store');
+            Route::get('/admin/category-image/{categoryImageId}/delete', 'delete');
         });
     });
 });
+    
     // ROUTING group for reserved users only users who have pending reservation have access here
     Route::group(['middleware' => ['auth','isReserve']], function () {
         Route::get('/reserve/wait',[ReservationController::class,'waiting'])->name('reserve.wait');
@@ -77,11 +75,15 @@ Route::middleware(['auth','verified'])->group( function(){
         Route::get('/reserve/{id}/{apartment}/{reservation}/update',[ReservationController::class,'update'])->name('reserve.update');
     });
     Route::group(['middleware' => ['auth','isRenter']], function () {
-        Route::get('/renters/index',[RenterController::class,'index'])->name('renters.index');
+        Route::controller(SubmitReportController::class)->group(function() {
+            Route::get('/renters/home','home')->name('renters.home');
+            Route::get('/renters/report/index','index')->name('renters.report.index');
+        });
     });
 
-    Route::get('/reserve/{apartment}/index',[ReservationController::class,'index'])->name('reserve.index')->middleware('auth');
-    Route::post('/reserve/create',[ReservationController::class,'create'])->name('reserve.create');
+
+Route::get('/reserve/{apartment}/index',[ReservationController::class,'index'])->name('reserve.index')->middleware('auth');
+Route::post('/reserve/create',[ReservationController::class,'create'])->name('reserve.create');
 
 
 
@@ -106,11 +108,12 @@ Route::middleware(['auth','verified'])->group( function(){
 // route to see the full details for a certain apartment
 Route::get('visitors/{apartment}/detail',[VisitorPageController::class,'display'])->name('visitors.display');
 
-
-Route::get('/available', function () {
-    return view('available');
-})->name('available');
-
+// all types of user that is currently logged in should have access in profile module to update or delete their profile
+Route::group(['middleware' => ['auth']], function () {
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+});
 
 
 require __DIR__.'/auth.php';
